requirements.txt
flask
Crawl4ai
mysql-connector-python
BeautifulSoup4
python-dotenv